requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
ffmpeg==1.4
ffprobe==0.5
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.1
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
youtube-dlc==2020.10.31
